requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-API==3.0.post1
Flask-RESTful==0.3.9
Flask-Script==2.0.6
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
